Install seturgent command
This has less utility than I thought because fvwm (at least) doesn't
seem to be able to handle more than one urgency hinted window (I'm not
sure the resolution of the issue but this is my guess).
Also seturgent doesn't unset the urgency when the window is focussed,
and fvwm (at least) only works on the rising edge.

So ultimately I can only use it once before having to close the window
to reset the urgency (or I'd have to manually unset the urgency which I
wouldn't remember to do).

xfce4-terminal is capable of setting the urgency flag itself, in
response to a bell ('\a') character.
Enable Preferences -> Advanced ->"Visual bell"
This may be what I'm after for notifications and in order to develop and
test custom handling of the urgency flag.

If I don't merge this into 'main', it's because I don't really need it.